refactor: Separate konva methods in a different file
