+++ user_interface/src/common/konvaMethods.js
@@ -1,0 +1,15 @@
+/**
+ * Given refs object of Vue Konva, awaits for v-stage to load
+ * @param {refs} vueKonvaRefs
+ * @param {number} intervalDuration
+ */
+export const waitForKonvaStageLoad = async (refs, intervalDuration) => {
+  return await new Promise((resolve) => {
+    const interval = setInterval(() => {
+      if (refs.stage !== undefined) {
+        resolve('ready');
+        clearInterval(interval);
+      }
+    }, intervalDuration);
+  });
+};
