+++ user_interface/src/common/utils.js
@@ -18,19 +18,3 @@ export const downloadURI = (uri, name) => {
   link.click();
   document.body.removeChild(link);
 };
-
-/**
- * Given refs object of Vue Konva, awaits for v-stage to load
- * @param {refs} vueKonvaRefs
- * @param {number} intervalDuration
- */
-export const waitForKonvaStageLoad = async (refs, intervalDuration) => {
-  return await new Promise((resolve) => {
-    const interval = setInterval(() => {
-      if (refs.stage !== undefined) {
-        resolve('ready');
-        clearInterval(interval);
-      }
-    }, intervalDuration);
-  });
-};
